finish

+++ source/game_sa/MenuManager.h
@@ -376,7 +376,7 @@ public:
     bool CheckMissionPackValidMenu();
     void CheckCodesForControls(eControllerType type);
 
-    int32 DisplaySlider(float x, float y, float h1, float h2, float length, float value, float spacing);
+    int32 DisplaySlider(float x, float y, float h1, float h2, float length, float value, int32 spacing);
 
     void DisplayHelperText(const char* key);
     void SetHelperText(eHelperText messageId);
